+++ docs/Changelog.md
@@ -1,3 +1,11 @@
 # Changelog
 
-- 20222-02-23: v0.1.3, support arxiv, googlescholar and scihub.
+**2022-02-23**
+
+- Version: v0.1.3,
+- Update: support arxiv, googlescholar and scihub.
+
+**2022-02-25**
+
+- Version: v0.1.4,
+- Update: Some improvements, like adding exe auto building and the progress bar.